requirements.txt
Django==5.2
djangorestframework==3.15.1
drf-spectacular==0.27.1
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.1
django-prometheus==2.3.1
prometheus-client==0.19.0
requests==2.31.0
urllib3==1.26.18
elasticsearch==7.17.0
python-logstash==0.4.6
celery==5.3.4
torch==2.0.1
transformers==4.55.0